+++ symbolic-ppdb/src/format/mod.rs
@@ -1,6 +1,7 @@
 mod metadata;
 mod raw;
 mod sequence_points;
+mod sourcelinks;
 mod streams;
 mod utils;
 
@@ -13,8 +14,10 @@ use watto::Pod;
 use symbolic_common::{DebugId, Language, Uuid};
 
 use metadata::{
-    CustomDebugInformation, CustomDebugInformationIterator, MetadataStream, Table, TableType,
+    CustomDebugInformation, CustomDebugInformationIterator, CustomDebugInformationTag,
+    MetadataStream, Table, TableType,
 };
+use sourcelinks::SourceLinkMappings;
 use streams::{BlobStream, GuidStream, PdbStream, StringStream, UsStream};
 
 /// The kind of a [`FormatError`].
@@ -101,6 +104,9 @@ pub enum FormatErrorKind {
     /// Tried to read contents of a blob in an unknown format.
     #[error("invalid blob format {0}")]
     InvalidBlobFormat(u32),
+    /// Failed to parse Source Link JSON
+    #[error("invalid source link JSON")]
+    InvalidSourceLinkJson,
 }
 
 /// An error encountered while parsing a [`PortablePdb`] file.
@@ -160,6 +166,8 @@ pub struct PortablePdb<'data> {
     blob_stream: Option<BlobStream<'data>>,
     /// The file's #GUID stream, if it exists.
     guid_stream: Option<GuidStream<'data>>,
+    /// Source link mappings
+    source_link_mappings: SourceLinkMappings,
 }
 
 impl fmt::Debug for PortablePdb<'_> {
@@ -229,6 +237,7 @@ impl<'data> PortablePdb<'data> {
             us_stream: None,
             blob_stream: None,
             guid_stream: None,
+            source_link_mappings: SourceLinkMappings::default(),
         };
 
         let mut metadata_stream = None;
@@ -280,6 +289,19 @@ impl<'data> PortablePdb<'data> {
                     .map_or([0; 64], |s| s.referenced_table_sizes),
             )?)
         }
+
+        // Read source link mappings.
+        // https://github.com/dotnet/runtime/blob/main/docs/design/specs/PortablePdb-Metadata.md#source-link-c-and-vb-compilers
+        const SOURCE_LINK_KIND: Uuid = uuid::uuid!("CC110556-A091-4D38-9FEC-25AB9A351A6A");
+        let mut source_link_mappings = Vec::new();
+        for cdi in CustomDebugInformationIterator::new(&result, SOURCE_LINK_KIND)? {
+            let cdi = cdi?;
+            // Note: only handle module #1 (do we actually handle multiple modules in any way??)
+            if let (CustomDebugInformationTag::Module, 1) = (cdi.tag, cdi.value) {
+                source_link_mappings.push(result.get_blob(cdi.blob)?);
+            }
+        }
+        result.source_link_mappings = SourceLinkMappings::new(source_link_mappings)?;
 
         Ok(result)
     }
@@ -363,6 +385,16 @@ impl<'data> PortablePdb<'data> {
     pub fn get_embedded_sources(&self) -> Result<EmbeddedSourceIterator<'_, 'data>, FormatError> {
         EmbeddedSourceIterator::new(self)
     }
+
+    /// Tries to resolve given document as a source link (URL).
+    /// Make sure to try [Self::get_embedded_sources] first when looking for a source file, because
+    /// function may return a link that actually doesn't exist (e.g. file is in .gitignore).
+    /// In that case, it's usually the case that the file is embedded in the PPDB instead.
+    pub fn get_source_link(&self, document: &Document) -> Option<Cow<'_, str>> {
+        self.source_link_mappings
+            .resolve(&document.name)
+            .map(Cow::Owned)
+    }
 }
 
 /// Represents a source file that is referenced by this PDB.
@@ -408,7 +440,7 @@ impl<'object, 'data> Iterator for EmbeddedSourceIterator<'object, 'data> {
             match row {
                 Err(e) => return Some(Err(e)),
                 Ok(info) => {
-                    if let metadata::CustomDebugInformationTag::Document = info.tag {
+                    if let CustomDebugInformationTag::Document = info.tag {
                         return Some(self.get_source(info));
                     }
                 }